+++ tests/test_base_module.py
@@ -77,9 +77,21 @@ def test_subscribe(base_module, mock_connection):
     module_mock.subscribe.assert_called_with("/test")
 
 
+def test_subscribe_str(base_module, mock_connection):
+    module_mock = mock_connection.return_value.awgModule.return_value
+    base_module.subscribe(base_module.test.node_info.path)
+    module_mock.subscribe.assert_called_with("/test")
+
+
 def test_unsubscribe(base_module, mock_connection):
     module_mock = mock_connection.return_value.awgModule.return_value
     base_module.unsubscribe(base_module.test)
+    module_mock.unsubscribe.assert_called_with("/test")
+
+
+def test_unsubscribe_str(base_module, mock_connection):
+    module_mock = mock_connection.return_value.awgModule.return_value
+    base_module.unsubscribe(base_module.test.node_info.path)
     module_mock.unsubscribe.assert_called_with("/test")